refactor: PostingForm 스타일 적용 코드를 다른 파일로 분리함
