+++ front-end/legeno-around-here/src/components/pages/posting/PostingForm.js
@@ -1,5 +1,4 @@
 import React, { useState, useEffect } from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 import Loading from '../../Loading';
 import TextField from '@material-ui/core/TextField';
 import IconButton from '@material-ui/core/IconButton';
@@ -14,58 +13,12 @@ import {
   ListItemText,
   Divider,
 } from '@material-ui/core';
-import { createPost } from '../../api/API';
-import { getAccessTokenFromCookie } from '../../../util/TokenUtils';
-import { findAllSectors } from '../../api/API';
 import AddPhotoAlternateIcon from '@material-ui/icons/AddPhotoAlternate';
-import SectorApplyButton from '../sector/SectorApplyButton';
 
-const useStyles = makeStyles((theme) => ({
-  grow: {
-    flexGrow: 1,
-  },
-  menuButton: {
-    marginRight: theme.spacing(0),
-  },
-  title: {
-    display: 'block',
-  },
-  sectionDesktop: {
-    display: 'flex',
-  },
-  modal: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  paper: {
-    backgroundColor: theme.palette.background.paper,
-    border: '2px solid #000',
-    boxShadow: theme.shadows[5],
-    padding: theme.spacing(2, 4, 3),
-  },
-  list: {
-    height: 400,
-    overflow: 'auto',
-  },
-  uploadPhoto: {
-    opacity: 0,
-    position: 'absolute',
-    zIndex: -1,
-    pointerEvents: 'none',
-  },
-  selectSectorButton: {
-    display: 'inlineBlock',
-    fontSize: '140%',
-    color: '#3366bb',
-  },
-  sector: {
-    display: 'inline',
-    borderRadius: 100,
-    backgroundColor: 'skyblue',
-    padding: '5px 8px 5px 8px',
-  },
-}));
+import { createPost, findAllSectors } from '../../api/API';
+import { getAccessTokenFromCookie } from '../../../util/TokenUtils';
+import useStyles from './PostingFormStyles';
+import SectorApplyButton from '../sector/SectorApplyButton';
 
 const PostingForm = () => {
   const classes = useStyles();
